Eliminate redundancy; start adding tests
