+++ graylog2-server/src/test/resources/org/graylog2/plugin/inputs/transports/util/server.key.pem.pkcs8.protected
@@ -1,0 +1,30 @@
+-----BEGIN ENCRYPTED PRIVATE KEY-----
+MIIFKzBVBgkqhkiG9w0BBQ0wSDAnBgkqhkiG9w0BBQwwGgQUHbrDlg6Kts1XnG/K
+sVDnICXPlG8CAggAMB0GCWCGSAFlAwQBKgQQLAdrno8pm9eMJR+LgFzIzwSCBNAt
+AgdfIzMnDQB2J+WmrxPR6ajhA3KlUBQdC5Bh4SJ3QVAk2L6f8ugLkZQYGB9OIhi0
+vGKHKPv4BF+ociFV2UGnDuygvQq4N9mqWV88LBJQ3sBqQFVKNVMxRl34OOf85apy
+x06E/+LrRTXihg3WqCX6yVrAkekua1oFCcsKlsqtzffECoxaMR5jJ7VDIyHdmb80
+/hGDfR4Bk62wvl2FZ8Rs2sVsccLOMPDvp/9Vb4nKXYl6ts1YTpQBWgCU3KYW4oLJ
+DeZ4ikgIIsF3JEFSRLvrQ6Zm07MUmnUYWW7qjqUCPm9UmYGnNnRTwvMCQ7Q9sEM3
+GX0vsFp9ucGO2W42/qpM97+uiZfCLNUty2gQGNB9jjbT6agQ0SDFSivdeDLu9VMD
+iXTJdS/zTXnGq1/Kfz8Dk4OXiaFNk8ceOAIWRjCxgP4fn2lL+4iLax9Vu2z84rv2
+Raor0RxZZSR8AkhhrdFw8B/pDGYKZXtr9NT6wJ4oLtWyjYhGJlHxSNWzJTWalhe3
+Q9myoqCW2aBiOq65jTuLt6mmBLfXO4P04FmgEu2QQ+m1Zdn54FC6vh14/6xjh4lr
+I/Ff8Y/86pcvdyW1drIQ8Gzmy6sxqxWaAxH9jJmw+IaiR36/xseh1B8qtWbma/fE
+68dMDYpCuEhe0UyJFrI7abR98JwN9k3uIRcdTsXlkXko8I+83PUwG78Er51jguF1
+C8MVhu5Q53qSpHDfgDk/cPIJKCa2h381BX93q6FSbkdskMOVq8QjUp4Rgou1loCH
+r4bWpnM42QARYKksbinUzfO7arleRr3w4ov/+YQPZymqkUmjSIVf0UAqKaM8yRzu
+YKfttLimsELVEeOBma4WFX4GnRnBh/NXgHDWCNa3ZqLDuKRR2z2lNKbpIBhuULxQ
+y0Q8jSlAvNhBLaM8iAsHQMS5mr9WVJNt8e0E7T4ttLswGo8PqQzqdiGl6nxqaXBP
+kbjI42xGymUu/K9uY/jLVTB/U7FrvIEf5qWa5uHvRUW7KnsiCzKSDniCh7QTWGas
+mZIfeGIMhzxybSBYd0Mh6Sv9fS85VEro6Zhijwj5Uh4vmIlshJPO96lIOWaeYQ30
+22CZhvYxCuZhPIx4EfrripFdM8L4+kqaIGs2QtGXCflXt93aQioOfdr+y4vS5UJg
+lc4VdzAAdsts0UfvRbVaX7cdP1TfzgvmNaX+PbftBaF1VhwlXqjdMjYYftO6iM4w
+SdM8fHaV2piSd2TKVNO824GoIfngh6mrMDAVWajfuHkgoltqLmQFZoQ1bqLc6C9R
+TzfbTjjLce30npKivpubiWxbY4B4k6vlRrl3P/7y3uf8yVfn76/oK0vyIG/SSeeh
+XTwaQ6aVYPcmE/OhTG58mKIQmFg6p/a4B07+Cg4fdshe0WP7EKuEotaCvcCVzaXf
+6OxCuj431Wr7gZsGEKpixW5NxMIiv0nyoY27OmQ9j1TOfWAtKS55hte638nc5sbP
+CzzWoMLadVLfKFgl4exUPp1IKHrG13FfTDtOv8tLP7Io6lvXg6LoONB4YgjhB6Mr
+PNwnGqqXQalR3tnPDHyGOO0NohRNedVjAWW7c8MGhc8Ka+VEdHjyVxHr6lL3LGsc
+whVxfoYNaaDr3vh2wxxlqoHzCvH0eBm6CA1bu0d8Cw==
+-----END ENCRYPTED PRIVATE KEY-----
