+++ graylog2-server/src/test/resources/org/graylog2/plugin/inputs/transports/util/server.key.pem.pkcs8.unprotected
@@ -1,0 +1,24 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCbx/DAO51APqVzEm4OSfhopldS
+Tde5i1znQPPUjDxsSiLarR4RhH7RI1N9BXDOsGCQOBSwBkQvM+o+uOLXUxAU2SlBQ+YXI6zoNp+r
+a6QtS3hq8FE6Umj+G4o49OW4BsU1UKntJtBKBhPD+Q4Fl6pYCVm5naIS6JK2bt8DgyJhJLOjgLQT
+wKhbgZo/akAAIzF9NT1IXbR4wa9xq7sDn7RyTPDHj7GAbmaThWUehit/HHLFk4eqnSwAhN/YkBmp
+WDMuNp/KDmoW8EVDMZu10QQi8zcdGUmV5ocKHZKazKXFrVGa+m7/f2l6flevBX3ZQZ+gC4h1Mwxr
+s1qoImo3i6T1AgMBAAECggEBAIJZfRP9TgmR4fM1uq1uK9O1TF/sDtSMzoIakWyKo4yhRCTissRJ
+r6vj+VolkU0BvOswYe8gbzFoKVuubPFr6YroY2LfI8q86VOtUVGo0kmsWuhwiidZCc2G0dPy571i
+vOVtx6JO1f76stNKtPbs/3aG7HMqeZGE1i6c+zL/KWdKdj8Is0hCr8mgBeHdY2XiWr3OK6yTBg64
+AJDgwv4sdRuWvscT6biCJyvwTLH/M7Iwhg6s2e065Z6eV47v3anXsSgnEwgAoPa9h1f2INt8JscB
+cLVwAcIWkW4v2zyabuLblrbDMi58o5DyZlv1ASHe7Am1UqqS2OswvRXgBBwYO8kCgYEAzila6tyq
+PWTlxMLTxWofVMsknqdx4W0rbnyL+O0PGAautV1ob93NZbOtQ8ajr99ZfHzNjXlD1aHT84rwMBNg
+xypYCU7FtN6RXJxEj9Hcik4XjAkX383xlNsRjnolVJ6xn4P2eZWCNaQlGsaX1m7gShOdH+9wBRFr
+zpuAus7blIsCgYEAwXC0Mg+E7O1dy2m7NmK2MCUwrxXlZdrjEbg3HA0Zo2A8yzCNfQOXCvK5rj/v
+04pOGEOMybK45xulLULhhJMYBUkLD9MAvKeyrSI31osAJTorHVqbAP7+l81S/HqOoijsSG92H0C1
+SrRdEbSyfzVF/z9cVxnVRlixoM09hbYRXH8CgYEAr5qcPbduLtNVv66wEdTeESzj6Vv5bam+t+LQ
+m1zZYX63q/Z/ceH9Babxa+y0Gqdlotrvf5UOSdcJIcZz5o9q/Nq9mFgHoZN/P+I/qD163De+aOKJ
++nGCWGftYW37hEOUSqpt9Ix8F9LUZCVKyjUxSQfB66TYxwhayua71ONEOBcCgYAzeSkJ9SU64oLF
+4PGnLCcssoi83ke3jcJEcc+AUGAVIEOtEdNkg7MKf8I3At5vXOsg+8CX303ieXG/lH9sBrQqm+As
+qsoUaU9bULKpDT3RqiPU4MpEPpZCkKlXc8lGjzsonzhAYribE1l8n6erv+JLZHHO35nn6SwDDPok
+/WirQwKBgFrLyESdc05c1/bgf6t3dewLVbVU4EgUEbt35vFPi8N8C5DdkjTYrXyN6lWYhbhCJ5Wd
+R/n3YnWOHSROE/2tvkiXzh3crFr8c8R8X1nhtwj3XvseZC2XrY5ujN0SFV3bdup3dQ8U03ncHIpC
+OTJ4pnrl9h2oxsyjPDElXAv5+cxu
+-----END PRIVATE KEY-----
